Properly handled ShapedOreRecipes
This code is pretty awful, but I’m aiming to get everything working perfectly first and refine my methods of doing so later. I’m probably going to make this configurable in some way, as it would be a disaster if a mod registered an OreDictionary entry for cobblestone or anything like that. Up next, ShapelessOreRecipes!